+++ src/main/java/com/care/dao/MemberDAO.java
@@ -9,11 +9,11 @@ public interface MemberDAO extends DAO{
 
     Member getMember(String email) throws SQLException;
 
-    Member getMember(int memberId) throws SQLException;
+    Member getMember(long memberId) throws SQLException;
 
     int addMember(Member member) throws SQLException;
 
-    int editMember(int memberId, Member member) throws SQLException;
+    int editMember(long memberId, Member member) throws SQLException;
 
-    int setMemberStatus(int memberId, Status status) throws SQLException;
+    int setMemberStatus(long memberId, Status status) throws SQLException;
 }
